+++ src/main/java/jointfaas/serverless/Request.java
@@ -1,4 +1,4 @@
-package org.example.serverless;
+package jointfaas.serverless;
 
 import java.io.IOException;
 import java.io.InputStream;
