+++ src/main/java/jointfaas/serverless/Worker.java
@@ -1,11 +1,10 @@
-package org.example.serverless;
+package jointfaas.serverless;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.lang.reflect.InvocationTargetException;
-import java.nio.ByteBuffer;
 
 public class Worker implements Runnable {
